refactor: eject from monorepo to single package structure
