--- lerna.json
@@ -1,12 +1,0 @@
-{
-  "command": {
-    "publish": {
-      "allowBranch": "release/*",
-      "conventionalCommits": true
-    }
-  },
-  "packages": [
-    "packages/*"
-  ],
-  "version": "2.0.0-pre.5"
-}
