--- packages/dom-element-to-react/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "dom-element-to-react",
-  "version": "0.1.0",
-  "description": "Convert the children of a node to React elements",
-  "main": "lib/index.js",
-  "author": "Measured Co. <hello@measured.co>",
-  "license": "MIT",
-  "scripts": {
-    "compile": "tsc"
-  },
-  "peerDependencies": {
-    "react": "^16.4.0"
-  }
-}
