--- packages/dom-element-to-react/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-
-  "compilerOptions": {
-    "outDir": "lib"
-  },
-
-  "include": ["src/**/*"],
-  "exclude": ["src/**/__tests__"]
-}
