--- packages/react-from-markup/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "react-from-html",
-  "version": "0.1.0",
-  "description": "Convert regions of markup into React, with custom callbacks",
-  "main": "lib/index.js",
-  "author": "Measured Co. <hello@measured.co>",
-  "license": "MIT",
-  "scripts": {
-    "compile": "tsc"
-  },
-  "dependencies": {
-    "dom-element-to-react": "^0.1.0"
-  },
-  "peerDependencies": {
-    "react": "^16.4.0",
-    "react-dom": "^16.4.0"
-  }
-}
