+++ src/IRehydrator.ts
