+++ src/dom-element-to-react/IAttributeList.ts
