+++ src/dom-element-to-react/attributes/index.ts
