+++ src/dom-element-to-react/attributes/nameMap.ts
