+++ src/dom-element-to-react/specialElementHandlers.ts
