+++ src/dom-element-to-react/staticToReactElement.ts
